Access control: Added candidate policy

+++ app/Domains/Candidates/Http/Controllers/CandidateController.php
@@ -9,7 +9,8 @@ use Illuminate\Validation\ValidationException;
 use Candidatozz\Support\Http\Controllers\Controller;
 use Candidatozz\Support\Database\Repository\ModelNotFoundException;
 use Candidatozz\Domains\Candidates\Contracts\CandidateServiceContract;
-use Candidatozz\Domains\Candidates\Transformers\CandidateTransform;
+use Candidatozz\Domains\Candidates\Transformers\CandidateTransformer;
+use Candidatozz\Domains\Candidates\Models\Candidate;
 
 class CandidateController extends Controller
 {
@@ -38,7 +39,9 @@ class CandidateController extends Controller
      */
     public function index()
     {
-        return $this->response()->collection($this->candidateService->paginate(), new CandidateTransform);
+        $this->authorize('index', Candidate::class);
+
+        return $this->response()->collection($this->candidateService->paginate(), new CandidateTransformer);
     }
 
     /**
@@ -49,32 +52,35 @@ class CandidateController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create', Candidate::class);
+
         try {
             $this->validate($request, [
                 'first_name' => 'required',
                 'last_name' => 'required',
-                'email' => 'required|email',
+                'email' => 'required|email|unique:candidates',
                 'gender' => 'required',
-                'curriculum_vitae' => 'required|mimes:doc,docx,pdf',
             ],[
                 'first_name.required' => 'Nome obrigatório.',
                 'last_name.required' => 'Sobrenome obrigatório.',
                 'email.required' => 'E-mail é obrigatório.',
                 'email.email' => 'E-mail inválido.',
+                'email.unique' => 'E-mail já encontra-se cadastrado.',
                 'gender.required' => 'Sexo é obrigatório.',
-                'curriculum_vitae.required' => 'Por favor envie seu currículo.',
-                'curriculum_vitae.mimes' => 'Só é aceito currívulos nos formatos doc, docx ou pdf.',
             ]);
 
-            $candidate = $this->candidateService->create($request->all());
-            $this->candidateService->saveCurriculum($request->file('curriculum_vitae'), $candidate->id);
+            $candidate = $this->candidateService->create($request->except('curriculum_vitae'));
+
+            if ($request->file('curriculum_vitae')) {
+                $this->candidateService->saveCurriculum($request->file('curriculum_vitae'), $candidate->id);
+            }
 
             return $this->response()->withSuccess('Candidato criado com sucesso');
 
         } catch (ValidationException $e) {
             return $this->response()->withUnprocessableEntity($e->errors());
         } catch (Exception $e) {
-            return $this->response()->withError('Ocorreu um erro ao criar o candidato');
+            return $this->response()->withError('Ocorreu um erro ao criar o candidato' . $e->getMessage());
         }
     }
 
@@ -86,9 +92,12 @@ class CandidateController extends Controller
      */
     public function show($id)
     {
+        $candidate = $this->candidateService->find($id);
+        $this->authorize('show', $candidate);
+
         try {
 
-            return $this->response()->item($this->candidateService->find($id), new CandidateTransform);
+            return $this->response()->item($candidate, new CandidateTransformer);
 
         } catch (ModelNotFoundException $e) {
             return $this->response()->withError($e->getMessage());
@@ -106,6 +115,8 @@ class CandidateController extends Controller
      */
     public function update(Request $request, $id)
     {
+        $this->authorize('update', Candidate::class);
+
         try {
             $this->validate($request, [
                 'first_name' => 'required',
@@ -113,7 +124,7 @@ class CandidateController extends Controller
                 'email' => 'required|email',
                 'gender' => 'required',
                 'birth_date' => 'required|before:16 years ago',
-                'curriculum_vitae' => 'required_if:has_curriculum_vitae,false|sometimes|mimes:doc,docx,pdf',
+                'curriculum_vitae' => 'required_if:has_curriculum_vitae,false|mimes:doc,docx,pdf',
             ],[
                 'first_name.required' => 'Nome obrigatório.',
                 'last_name.required' => 'Sobrenome obrigatório.',
@@ -122,7 +133,7 @@ class CandidateController extends Controller
                 'gender.required' => 'Sexo é obrigatório.',
                 'birth_date.required' => 'Data de nascimento é obrigatório.',
                 'birth_date.before' => 'O candidato precisa ter no mínimo 16 anos.',
-                'curriculum_vitae.required_if' => 'Por favor envie seu currículo.',
+                'curriculum_vitae.required_if' => 'Por favor, envie o currículo.',
                 'curriculum_vitae.mimes' => 'Só é aceito currívulos nos formatos doc, docx ou pdf.',
             ]);
 
@@ -151,6 +162,8 @@ class CandidateController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('delete', Candidate::class);
+
         try {
 
             $candidate = $this->candidateService->delete($id);
@@ -171,10 +184,11 @@ class CandidateController extends Controller
      */
     public function curriculumDownload($id)
     {
-        try {
+        $candidate  = $this->candidateService->find($id);
+        $this->authorize('curriculum-download', $candidate);
 
-            $candidate  = $this->candidateService->find($id);
-            $path       = Storage::path($candidate->curriculum_vitae);
+        try {
+            $path = Storage::path($candidate->curriculum_vitae);
 
             return response()->download($path);
